feat: install npm packages and settings eslint and webpack configs
