+++ .eslintrc.json
@@ -1,0 +1,38 @@
+{
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": [
+    "plugin:prettier/recommended",
+    "eslint:recommended",
+    "plugin:node/recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier"
+  ],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module",
+    "project": "./tsconfig.json"
+  },
+  "plugins": [
+    "@typescript-eslint"
+  ],
+  "root": true,
+  "rules": {
+    "no-debugger": "off",
+    "no-console": 0,
+    "class-methods-use-this": "off",
+    "import/extensions": "off",
+    "@typescript-eslint/no-explicit-any": "error",
+    "semi-style": ["error", "last"],
+    "node/no-unsupported-features/es-syntax":"off",
+    "node/no-missing-import": "off"
+    //["error", { "allowModules": ["types.d.ts"] }]
+  },
+  "ignorePatterns": [
+    "**/*.js",
+    "dist/"
+  ]
+}
